split no_mutabble_objects to module
