+++ flake8_elegant_objects/no_mutable_objects/__init__.py
@@ -1,0 +1,5 @@
+"""No mutable objects principle checker module."""
+
+from .core import NoMutableObjects
+
+__all__ = ["NoMutableObjects"]
