+++ flake8_elegant_objects/no_mutable_objects/contract_checker.py
@@ -1,0 +1,58 @@
+"""Immutability contract checker for proper immutability declarations."""
+
+import ast
+
+from ..base import ErrorCodes, Violations, violation
+
+
+class ImmutabilityContractChecker:
+    """Checks that classes properly declare and enforce immutability contracts."""
+
+    def check_immutability_contract(self, node: ast.ClassDef) -> Violations:
+        """Check if class has proper immutability declarations."""
+        violations = []
+
+        has_slots = any(
+            isinstance(item, ast.Assign)
+            and any(
+                isinstance(target, ast.Name) and target.id == "__slots__"
+                for target in item.targets
+            )
+            for item in node.body
+        )
+
+        has_setattr_override = any(
+            isinstance(item, ast.FunctionDef) and item.name == "__setattr__"
+            for item in node.body
+        )
+
+        has_mutable_attrs = False
+        for item in node.body:
+            if isinstance(item, ast.FunctionDef) and item.name == "__init__":
+                for stmt in ast.walk(item):
+                    if isinstance(stmt, ast.Assign):
+                        for target in stmt.targets:
+                            if (
+                                isinstance(target, ast.Attribute)
+                                and isinstance(target.value, ast.Name)
+                                and target.value.id == "self"
+                                and self._is_mutable_value(stmt.value)
+                            ):
+                                has_mutable_attrs = True
+                                break
+
+        if has_mutable_attrs and not (has_slots or has_setattr_override):
+            violations.extend(
+                violation(
+                    node,
+                    ErrorCodes.EO024.format(
+                        name=f"class {node.name} with mutable attributes but no immutability enforcement"
+                    ),
+                )
+            )
+
+        return violations
+
+    def _is_mutable_value(self, node: ast.AST) -> bool:
+        """Check if a value is mutable."""
+        return isinstance(node, ast.List | ast.Dict | ast.Set)
